Rename PubDate to CommentTimestamp
