+++ client/coral-embed-stream/src/components/Comment.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import TagLabel from 'talk-plugin-tag-label/TagLabel';
-import PubDate from 'coral-framework/components/PubDate';
+import CommentTimestamp from 'coral-framework/components/CommentTimestamp';
 import {ReplyBox, ReplyButton} from 'talk-plugin-replies';
 import {FlagComment} from 'talk-plugin-flags';
 import {can} from 'coral-framework/services/perms';
@@ -467,7 +467,7 @@ export default class Comment extends React.Component {
               <span className={`${styles.bylineSecondary} talk-stream-comment-user-byline`} >
                 <Slot
                   fill="commentTimestamp"
-                  defaultComponent={PubDate}
+                  defaultComponent={CommentTimestamp}
                   className={'talk-stream-comment-published-date'}
                   created_at={comment.created_at}
                   queryData={queryData}
