+++ client/coral-framework/components/CommentTimestamp.css
@@ -1,4 +1,4 @@
-.pubdate {
+.timestamp {
   display: inline-block;
   color: #696969;
   font-size: 12px;
