+++ client/coral-framework/components/CommentTimestamp.js
@@ -2,17 +2,17 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import {timeago} from 'coral-framework/services/i18n';
 import cn from 'classnames';
-import styles from './PubDate.css';
+import styles from './CommentTimestamp.css';
 
-const PubDate = ({className, created_at}) => (
-  <div className={cn(className, styles.pubdate, 'talk-comment-pubdate')}>
+const CommentTimestamp = ({className, created_at}) => (
+  <div className={cn(className, styles.timestamp, 'talk-comment-timestamp')}>
     {timeago(created_at)}
   </div>
 );
 
-PubDate.propTypes = {
+CommentTimestamp.propTypes = {
   className: PropTypes.string,
   created_at: PropTypes.string,
 };
 
-export default PubDate;
+export default CommentTimestamp;
