+++ plugin-api/beta/client/components/index.js
@@ -3,4 +3,4 @@ export {default as ClickOutside} from 'coral-framework/components/ClickOutside';
 export {default as IfSlotIsEmpty} from 'coral-framework/components/IfSlotIsEmpty';
 export {default as IfSlotIsNotEmpty} from 'coral-framework/components/IfSlotIsNotEmpty';
 export {default as CommentAuthorName} from 'coral-framework/components/CommentAuthorName';
-export {default as PubDate} from 'coral-framework/components/PubDate';
+export {default as CommentTimestamp} from 'coral-framework/components/CommentTimestamp';
